feat: put synclistener in own class and corrected other issues mentioned by dart
Moved to version 3.1.0
