+++ lib/version.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 //Print version number
 void version() {
-  final String version = "3.2.0";
+  final String version = "3.1.0";
   stdout.writeln('Version : $version');
 }
